Move verifier modules into 'intellij-plugin-verifier' directory.
To open the project from clean installation it is enough to import the 'plugins-verifier-service' module, which will load all the other modules using Gradle Composite Builds.

Historically, this Git repository contains several projects. Modules of the verifier used to be in the root and secondary projects were in separate directories:
<root>
 /verifier-parent
 /verifier-cli
 /intellij-plugin-structure (separate project)
 /plugins-verifier-service (separate project)

It is not really convenient, so let's move all verifier's modules into a dedicated directory.