fix: skip Kotlin default methods leaking in inheritors
Kotlin's default method are compiled differently than Java's default
methods. In short Kotlin is compiling an inner class `DefaultImpls`
in the parent type. The class that extends or implement the interface
but not the _default_ method gets overridden anyway in the bytecode.

Unfortunately while this code was never written by the plugin author,
it can get flagged if the method signature used any internal types.

Should fix MP-5395.