Add plugin pattern possibility for -keep-only-problems filter
The change adds the opportunity to mention the pattern of the plugins for which the `keep only` filter can be applied.

The need for such change has appeared in testing the Kotlin plugin compatibility with Android Studio and bundled plugins. We want to keep only problems matching pattern .*kotlin.* for all plugins except Kotlin. because there may be issues with platform calls from the Kotlin plugin site.