Remove the ability to disable downloading of unavailable optional dependencies (#1314)
If a plugin declares an optional dependency on another plugin that isn’t available in the IDE, that dependency is downloaded from the JetBrains Marketplace.

This was in the Plugin Verifier as per original specs, but has been removed in the new dependency tree based resolution mechanism.

Remove ability to disable this dependency download mechanism, as there is no use-case for that.