Test that Coverage plugin in 243 does not have a transitive dependency on Kotlin
The `com.intellij.modules.lang` is an alias of IDEA CORE. It declares multiple inline CDATA modules. One of those is `intellij.kotlin.onboarding-promoter`.
This promoter in the inline CDATA says:
```
<dependencies>
    <plugin id="org.jetbrains.kotlin" />
    <module name="kotlin.features-trainer" />
    <module name="intellij.java.featuresTrainer" />
</dependencies>
```

The `plugin.xml` in `product.jar` does not declare any dependencies at all (neither v1, nor v2).

`IDEA CORE` is a plugin that has a set of content modules, each of those content modules load independently (given the IDEA CORE loads) with their own classloader. So `IDEA CORE` (as the implicit main module of the plugin) has no dependencies.

`intellij.kotlin.onboarding-promoter` is an optional content module that loads if its dependencies are available, and it also implicitly depends on the `IDEA CORE`.

There should be no transitive dependency on Kotlin from the `Coverage` plugin.